+++ src/rate.js
@@ -10,9 +10,9 @@
  */
 function rate ( obj, req, res, next ) {
 	var headers = ["x-ratelimit-limit", "x-ratelimit-remaining", "x-ratelimit-reset"],
-	    results = obj.rate( req ),
-	    valid   = results.shift(),
-	    config  = obj.server.config.rate;
+	    config  = obj.server.config.rate,
+	    results = obj.rate( req, config.override ),
+	    valid   = results.shift();
 
 	array.each( headers, function ( i, idx ) {
 		res.setHeader( i, results[idx] );